Fix number of columns, drop idle from charts
Currently lots of charts have the end_columns fields (on the hdrs
row) too big.
This change creats the header line for the chart and then we count
the number of columns in the header row. This is used as the end
column field.
So I had to build the header string before I do the hdrs row.
Also a handful of charts include the idle cpu time on them.
It is useful to have idle cputime on the charts but it is always
so much bigger than the rest of the values that I end up deleting
the idle line or column.
I got tired of then deleting the idle line or column so I just
don't put it on the chart now.
If you want to see it you can do 'idle_i = -1' instead.