Massive changes to use more human style equations
Currently the equations used are rpn notation or bc_eqn (regular
'1+2*(3+4)') style but the equaations are built by pushing the
operands and operators onto a list.
This makes it difficult to read.
I use the bc_eqn.awk script for the current bc equation support.
But this script (from rosetta) is difficult to debug and enhance.
So I wrote bc_eqn2.awk which seems pretty capable so far.
This commit changes about half of the equations to use bc_eqn2
so far.