replace rest of equations with bc_eqn2 format
The previous commits switched about half the equations with
bc_eqn2 format equations.
Now use bc_eqn2 for all the equations.
Also have a common got_lkfor_init() function to init common
fields.
Now we can also check if any of the events in the input perf_stat
file isn't used in a completed equation. By 'completed' eqn I
mean an equation that has all its dependencies met so that the
equation will be used. This check will tell me if there is a set
of events that need an equation added.