Add more ARM equations and remove some old code
ARM frequency (at least on graviton) seems to be fixed.
So read it from a file and pass it to script. Allow for equations
with constant value (like the arm_cpu_freq avg_freq equation).
I'm not quite sure how to calculate %not_halted on arm with the
perf events. If the freq is fixed (in this case as 2.5 Ghz) then
I can calculate the %not_halted based on the pseudo code:
%not halted = perf cycles / (num_cpus * arm_cpu_freq * interval)

Also removed some old code.