Get arm_cpu_freq from tsc_freq.txt file
For the time being, assume arm freq is constant. I don't think it
is on phones but it seems to be on graviton servers.
I changed tsc_freq.c to emit cycles in addition to tsc_freq.
tsc_freq is doing a 'spin.x -w freq_sml' type loop where it
does an unrolled loop of add instructions (which run 1/cycle) to
get the cycle freq.
So if the perf data is for arm then get the arm_cpu_freq from the
tsc_freq.txt file and pass it to perf_stat_scatter.awk.