remove gawk dependcies and add getopts options
Currently there are some dependencies on gawk features:
The script does '@include' and strtonum().
Some linuxes only have an old mawk which dddoesn't support these
gawk features.
So I want to remove the need for gawk.
I added getopt so that we could pass in the args needed to figure
out the cpu code name.
Now the script works with gawk/awk/mawk.