More cleanup and chg to only do MSRs on Intel
Do a little more cleanup.
Currently reading and writing of MSRs is done on Intel and AMD and
the script will try to do MSRs on ARM too.
MSRs don't work on ARM.
The the purpose of reading/writing the MSRs in this script is to
display/control the turbo frequencies.
I really can only control the turbo frequencies on Intel.
On AMD the closest I can get is disable/enable turbo. This is not
really the same thing as limiting the turbo frequency on Intel.
By default currently if you do '-f allcore' or '-f reset' on AMD
then the script will disable or enable turbo.
The AMD behavior  is nowhere near the same thing as what happens
on Intel.
So add '-F 1' for AMD. If '-F 0' (the default) then no MSRs will
be read/written on AMD (so turbo won't be enabled/disabled).