Add more error checking and restrict changing
Be more carful about changing MSRs.
Only change to the -f allcore freq if we the current value of MSR
0x1ad is the expected after-bootup value.
Or else we won't know what to change it back to during -f reset.
Also be more careful about making sure we have identified the
chip and have the correct expected values.
Display an error message and set return code= 1 in case of errors.
Try to use case statements more to simplify code.