Allow multiple dirs for -d option and more phase
Frequently I want to have multiple dirs' data in the same xlsx so
that I can get charts all in one xlsx.
Currently the only way to get multiple dirs in the same xlsx was
to specify an upper level dir that has multiple data dirs below
it then each data dir will be put in the xlsx.
But lets say we have a disk test with phases that ramps from
1 thread->16 threads->64->128 threads.
Then we might want to specify that data dir multiple times and
select each phase (so we can see how the data changes as you ramp
up the load).
And maybe we have a fast and slow server and we want to see
compare performance of the fast and slow.
This change lets you specify a dir list like -d a:b:c where
a b and c is a data dir.
And you can specify a phase file like -P phase.txt
And -C phase_str for each dir like
-C phase0 -C phase1 -C phase2 where the data in dir 'a' will be
clipped to phase0, and dir data b will be clipped to phase1, etc.