Add clip to phase logic for a lot more data files
Many of the data files don't yet have clipping to a phase built
in. This change fixes that for lot of the files types.
Also add support for sys_gen_table_01.txt files (plot all the
data in the table, assume tsv type, and the 1st column must be
epoch ts, 2nd column relative ts.
Also add support for sys_20_proc_stats.txt (which is a subset of
the rd_infra stats.