Add --phase option. Fix ck options. Fix change new charts
Currently if you have multiple data dirs in the xlsx you only get
the charts for each data on a new line if there an 'desc.txt' in
the data dir and if the contents are different.
This isn't good for say where you are doing the same data dir
multiple times but with a phase for each dir instance.
So change the logic for 'start a new row of charts' such that
it will start a new row for a new dir.
Also the 'for opt in (a, b)' logic created false matches since
it aactually matches on substrings.
Some day I'll fix the duplicate code for checking the options.
Also add a '--phase' option which gets printed above each chart.