Fix typo leading to 1 in PID_LST and more timing
Currently there is a typo in building the list of PIDs to be
killed at the end of the script. It results in a PID of 1 being
added to the list (which is very bad).
Fix the typo.
Also add a display of the current elapsed time for each iteration.