add support for getting docker java symbol file
Currently the scripts can't see the the java symbols. We need to
get a symbol file from java procs in the containers. We use Gregg's
https://github.com/brendangregg/FlameGraph/blob/master/jmaps
and https://github.com/jvm-profiling-tools/perf-map-agent.
Bundled these into get_symbol_map_for_perf_from_java_in_container.tar.gz
So we have to check if the container has java. If so do a docker
cp of the tar.gz to the container /tmp dir.
Also need to know the mapping of the java host pid to the docker
pid. This is /proc/[host-pid]/status NSid: host_pid docker_pid
The perf script cmd needs the java symbol file so right before we
do 'perf script' we do a docker exec bash -c 'cd /tmp;./fg_jmaps.sh'
as same --user java_user (the same user as the owns the java in
the container).
So now we have the /tmp/perf-[docker_pid].map file for each java
in the container.
Now we need to copy the docker java symbol file to the host and
rename it with the host pid for the java process. We already know
the host->docker pid mapping. So do a:
docker cp [id]:/tmp/perf-[docker pid].map /tmp/perf-[host pid].map
Now we have to tell the perf script cmd that there are map files
in /tmp so add option --symfs /tmp to the perf script cmd.