Make getopts options a superset of perf_stat_scatter.sh
Trying to make bc_eqn3.sh a replacement for perf_stat_scatter.sh
The bc_eqn3.sh options have to be a superset of the ps
(perf_stat_scatter) options.
Not all ps options are as of yet implemented in bc_eqn3.
In particular, the clipping is not implemented, the reduction
stuff (drop x of y rows) is not, version stuff in the options
string are not implemented (like pxx stats).
But now the equations are all in the eqn_milan/arm/csx.txt files.
The processing takes about 2x longer so far.