Add ck if use eqn and allow not raise err for div by 0
Currently two functions have to get called to check the use_if_got
and skip_if_got conditions.
Fold those 2 into 1 function.
Also, the divide by 0 check sets an error code and the code does
some error handling (note that the div by 0 doesn't actually
happen).
But there are cases where the divisor will be 0. Such as if the
divisor is just an event and the event didn't get any samples.
So add an eqn file option that you can add to an equation
options: if_denom_zero_return_zero
Then, if we are going to div by 0, we don't set an error but the
equation is set to 0.