add support for tsv out_file
Currently this script just prints out the value for each equation
to stdout.
perf_stat_scatter creates a table of the perf data in a tsv.
The tsv can be read by tsv_2_xlsx.py to create a sheet in an xlsx
file with charts.
Most of the changes in this file are to add prt_to_out_file()
which implements most of the logic like perf_stat_scatter.