add options to make bc_eqn3 compat with perf_stat_scatter
The equations need to produce output compatible with
perf_stat_scatter so that the results can be compared.
Add lines like:
hdr_ps: string # for 'header perf stat'
   if hdr_ps is present then the out_file will use hdr_ps as
   the header (instead of the hdr: string.
hdr_alias: string  # another header string for the eqn
   This is needed because sometimes I need to compare to a metric
   say from intel perfspect tool (which calls the eqn hdr some
   other string). Using hdr_alias will produce 2 lines for the
   eqn, 1 with tag hdr_ps (or hdr), another with tag hdr_alias
hdr_alias_factor: number if this is present then the value for
   the hdr_alias is multiplied by the hdr_alias_factor
   This is useful since sometimes perfspect uses MB/s instead of
   GB/s
Rearrange equations so that they appear in the same order as
perf_stat_scatter (helps with matching output).
Add ice-lake topdown equations.
These equations are selected with the conditional:
options: skip_if_got{cpu_type==Ice Lake}
Add a charts option to indicate that an eqn goes into a chart.
charts: td_lvl1_chart
This option appears on lots of equations in the file but, in the
end, only equations that have all the events and satisfy the
conditionals are used in the chart.
Add an option to not raise an error if an equation tries to
divide by 0. For instance, the equation my be okay (all the
events are there, all the conditionals are satisfied) but the
count might be 0. If this 0 is the denominator then normally
bc_eqn3 raises an error and prints a bunch of debug info so you
can see what is going wrong.
But if we just have a zero count event then just return 0 for
the eqn. Use the equation option:
options: if_denom_zero_return_zero
I added a hdr_tag: option too for compatiblity but I don't
do anything with it yet.