Initial support for set_x_axis_date_axis
It would be good to support the x axis as a time series.
Currently the x axis is always just 'values' (that is not treated
as a time series). This usually is okay since I'm plotting every
X seconds. But if the data is not uniform then it won't show
properly.
This commit is partial support for set_x_axis_date_axis option
but it doesn't work yet.