Allow 'exit' to mark end of hostlist
Currently one can control the number of entries used from a host
list with the cmdline options -N x-y but sometimes I just want to
say "don't use the rest of the list".
If the only word on a line is 'exit' then I break out of the loop
building the host list and the rest of the file is ignored.
That is, if the line is:
exit
then the exit line and everything after it is ignored.