add titles and subtitles to flamegraphs and more
It can be hard to interpret the flamegraphs.
For each instance where we saw cgroup throttling we want to
produce a flamegraph of the call stacks. But, for each perf.dat
file, there is (maybe) a first interval with no throttling and
then a section (at the end of the file) with throttling.
isn't throttled
So I added a title and subtitle for the not throttled flamegraph
like:
  not throttled elapsed time covers 66.764 secs
  avg cpus/period= 0.136 vs cpu quota of 1.200 cpus/period
For the throttle flamegraph I added title & subtitle like:
  throttled elapsed time covers 1.099 secs
  throttled periods= 1 throttled_secs= 0.100. max= 1.182 cpus/period in interval time_09