Generate flamegraph even if no sigusr2 rec found
When the monitor script sees container throttling it sends sigusr2
to perf to dump the call stacks.
Then the extract script looks for sigusr2.
If there isn't any throttling then there won't be a sigusr2.
But there will be a sigint record where the sigint is sent to
shutdown perf.
This commit will permit generating flamegraphs for the case where
we didn't get any sigusr2 but we still have the sigint.
Currently it still generates 2 charts. Normally the 2nd chart
would be the throttling chart. It shows no throttling in the
flamegraph title but having the chart is confusing. I'll fix it
later.