allow processing all prf.dat files in list of dirs
Currently the script only does the prf.dat.txt files in 1 dir.
And it just produces 1 throttled/not_throttled svg per prf file.
This is okay but if you have lots of files in the dir then you
have to go through each svg.
And if you have lots of dirs then it becomes impossible.
The main part of this commit looks for throttled intervals and
puts the throttles calls stacks in 1 txt file (summar_thr.txt)
and puts the not throttled intervals in another summary_not.txt
file.
We also have to put stats (like total samples and elapsed time)
for each group of call stacks written to each summary file so we
can compute stats (like total elapsed time and cpus used/sec) in
the svg's titles.
The "what do we count as throttled" can be set at the cmd line
with a -p pxx (percent of quota... like -p 80 for 80% of cpu_quota
or -c cpus. Currently the 'interval' is hard coded to 0.1 seconds.
Note that the -o summary_file_prefix has to be passed to the
script to enable this "just create a summary throttled svg and a
not throttled svg" mode.
A list of prf.dat.txt files can be passed to the script for the
summary mode. The files should be sorted by the creation date
in the filename like
ls -1tr ${ARR[$i]}/prf_*.dat.*.txt | sed 's/dat./dat /' |sed 's/.txt/ txt/' |sort -nk 2 | sed 's/dat /dat./' | sed 's/ txt$/.txt/'
Assumes the files are named like prf_0.dat.2022030904372038.txt
where 0 is the container id.