add delay in creating prf.dat files with -m mode
Currently a prf.dat dump is triggered (via sigusr2) every time the
trigger condition is true. For the case of 'the container
throttled' this can be every 0.1 seconds. This can create too
many prf.dat files. Especially if I'm combining them anyway into
a summary svg... then I just look for cpu usage > xxx in a 0.1
second interval in the prf.dat file. So I don't need to a separate
prf.dat for each throttle trigger. I just scan the whole file for
cpu util > xxx.
This change only does sigusr2 at most every xx seconds where xx
is set by '-m trigger=xx'. The default is 60 seconds.
60 secs is (probably) not so long that the trigger interval will
get overwritten in the ring buffer. And it reduces the number of
prf.dat files... but it probably doesn't reduce the total MBs of
prf.dat files. Just (maybe) reduces the overhead of producing the
files a little and reduces the number of files (if there are lots
of triggers).
I probably should put in a "quit if total MBs of prf.dat files is
greater than xx MBs" to control the data produced.