Add more container stats, in/exclude certain stacks
Lots of little changes.
Add support for cpu, memory and blkio container stats.
The blkio stats seem to be pretty useless. It looks like the IO
gets done at the host level and the stats never make it to the
cgroup level. So although the blkio stats are non-zero sometimes
they change in the blkio stats when I monitor the container has
always been zero.
Added "include only stacks containing the string" and "exclude
all stacks containing the string" options. This is kind of for
debugging/sanity checking. I used it to show that the stacks
involved in throttling are indeed hitting the quota. That is, I
had garbage collection and a couple other stacks that consistently
showed up as throttling. I know GC is an issue but the other
stacks are usually (like 1 cpu vs need 12 cpus to throttle).
So I can exclude all the stacks for garbage collection and see if
the other 2 stacks still show up as throttling (and the they did).
Or I can select just those 2 stacks and then I see there 2 modes,
one mode is about 1 cpu and another mode is 12 cpus.
To do this include/exclude logic I had to, when reading the perf
.dat files, accumulate the whole sample+callstack and then check
if it should be included or excluded. If it is excluded then the
code 'disappears' that sample and its callstack so it doesn't
appear in any stats.
I also change the output file names from (the default):
summary_not.svg
summary_thr.svg
to
summary_not_thr.svg
summary_thr.svg