Use everything after -- as cmd to be run. RAM pwr
Currently if you want to run a cmd using the perf syntax:
 perf stat ... my_cmd my_cmds_arg1 [...]
then you have to use the -x my_cmd -X " my_cmds_arg ..." syntax.
This is cumbersome if the -X args have spaces and you have to pass
through multiple layers of scripts.
This commit lets you use the syntax:
./do_perf3.sh doperf3 options -- my_cmd my_cmds_arg my_cmds_arg2
So averything after the -- is assumed to be the command line you
want to run on the 'perf stat ... ' command.
Then the script uses the 1st arg after -- as the '-x arg' and
everthing after the 1st arg (if anything) as the -X args.

Also add support for the RAPL ram-energy event (if supported) to
get the DIMM power usage.